Improve error and disconnection handling.
All delegate methods are now optional. A new delegate method, -socketIoClient:didFailWithError:, is called when an underlying error occurs, or when the connection or heartbeat times out.

Removed tryAgainOnConnectTimeout as it is confusing and error-prone. Instead, delegates should check for SocketIoClientErrorConnectionTimeout and call -connect to try again if they desire. The queue of outgoing messages is preserved between attempts.

When heartbeat times out, explicitly disconnect the underlying socket immediately. Previously, the socket was left open so SocketIoClient could miraculously pop back open after a heartbeat timeout.