Retool SocketIoClient to be a bit more robust.
First, use a state variable instead of potentially contradictory boolean flags.

Introduce a new delegate method socketIoClient:connectDidFailWithError: that is called when an attempt to connect fails. Document that EITHER socketIoClient:connectDidFailWithError: is called, OR socketIoClient:didDisconnectWithError: is called, but not both. (The latter is called only if a connection is established, signaled by socketIoClientDidConnect:).

Make it safe to immediately retry the connection in the delegate error handling functions by deferring their execution until the next run loop.